requirements.txt
openai==1.12.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
networkx==3.1
matplotlib==3.8.2
tiktoken>=0.5.2,<0.6.0
faiss-cpu==1.7.4
numpy==1.24.3
pypdf==3.17.1
spacy==3.7.2
python-dotenv==1.0.0
unstructured==0.10.30
sentence-transformers==2.2.2
chromadb==0.4.22 